Reference variables, filters, and flywheels - Wolf
This adds reference variables named `speed` to FiltersOn (to control
filter speed), to FlywheelsForward(for flywheel forward speed), and to
FlywheelsBackward(for flywheel backward speed). They point to values in
the Shooter subsystem, but I thought it would be useful to add a
reference here as well. 

This also adds commands for the filters.